refactor(components): create common folder
