+++ src/app/(private)/lists/[id]/loading.tsx
@@ -1,5 +1,5 @@
 import { ContentWrapper } from '@/components/layout/content';
-import { ListDetailsViewSkeleton } from '@/components/list-details-view';
+import { ListDetailsViewSkeleton } from '@/components/common/list/list-details-view';
 
 export default function ListDetailLoading() {
   return (
