+++ src/app/(private)/lists/[id]/page.tsx
@@ -7,7 +7,7 @@ import {
 } from '@tanstack/react-query';
 
 import { ContentWrapper } from '@/components/layout/content';
-import { ListDetailsView } from '@/components/list-details-view';
+import { ListDetailsView } from '@/components/common/list/list-details-view';
 import { Card, CardContent } from '@/components/ui/card';
 import { ROUTES } from '@/constants/routes';
 import type { UserListViewDto } from '@/server/api/lists';
