+++ src/components/common/discover/discover-item-card.tsx
