+++ src/components/common/list/delete-list-dialog.tsx
