+++ src/components/common/list/list-card.tsx
@@ -7,10 +7,10 @@ import { ListVisibility } from '@generated/prisma/enums';
 import { useQuery } from '@tanstack/react-query';
 import { BookmarkIcon } from 'lucide-react';
 
-import { ItemsProgress } from '@/components/items-progress';
-import { ListEditButton } from '@/components/list-edit-button';
-import { ListSaveButton } from '@/components/list-save-button';
-import { ListTagsBadges } from '@/components/list-tags-badges';
+import { ItemsProgress } from '@/components/common/items-progress';
+import { ListEditButton } from '@/components/common/list/list-edit-button';
+import { ListSaveButton } from '@/components/common/list/list-save-button';
+import { ListTagsBadges } from '@/components/common/list/list-tags-badges';
 import { Body } from '@/components/typography/body';
 import { Caption } from '@/components/typography/caption';
 import { Label } from '@/components/typography/label';
